+++ CHANGELOG.rst
@@ -6,6 +6,8 @@ Version x.x.x
 
 - Unified command-line utilities into ``cnf``
 - Added ``cnf stats ...``
+- Added ``cnf simplify ...``
+- Added ``cnftools.simplify`` module
 
 Version 0.0.1
 -------------